+++ fs/sul/id/module.f.ts
@@ -10,7 +10,6 @@ import { toArray } from '../../types/list/module.f.ts'
 import {
     length,
     msb,
-    msbListToVec,
     uint,
     uintChunkList,
     unpack,
@@ -127,7 +126,7 @@ const hash2 = base32.compress(iv)
 const vecX20 = vec(0x20n)
 
 const hashMerge = (a: Id, b: Id): Id =>
-    hashId(uint(msbListToVec(hash2((asBase(a) << 0x100n) | asBase(b)).map(vecX20))))
+    hashId(uint(msb.listToVec(hash2((asBase(a) << 0x100n) | asBase(b)).map(vecX20))))
 
 const { concat } = msb